feat : added contact handlers and repo

+++ internal/domain/model/contact.go
@@ -6,9 +6,9 @@ import (
 
 type Contact struct {
 	gorm.Model
-	UserId         uint64 `gorm:"primaryKey;autoIncrement:false"`
+	UserId         uint   `gorm:"primaryKey;autoIncrement:false"`
 	User           User   `gorm:"foreignKey:UserId"`
-	ContactId      uint64 `gorm:"primaryKey;autoIncrement:false"`
+	ContactId      uint   `gorm:"primaryKey;autoIncrement:false"`
 	Contact        User   `gorm:"foreignKey:ContactId"`
 	ContactName    string `json:"contactname,omitempty"`
 	ShowNumber     bool   `json:"shownumber,omitempty"`
